+++ changelog.md
@@ -2,6 +2,7 @@
 
 # v1.0.1
 * Fixed the USB samplerate unlocking up to 768 kHz instead of 384 kHz
+* Fixed a guard condition when my other conflicting modules are disabled
 
 # v1.0.0
 * Initial release